+++ js/util.js
@@ -1,0 +1,19 @@
+// source link https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random
+
+const getRandomIntFromRange = (min, max) => {
+  if (min < 0 || max < 0 || max <= min) {
+    return null;
+  }
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+const getRandomFromRange = (min, max, decimalLength) => {
+  if (min < 0 || max < 0 || max <= min) {
+    return null;
+  }
+  return parseFloat((Math.random() * (max - min) + min).toFixed(decimalLength));
+};
+
+export {getRandomIntFromRange, getRandomFromRange};
